+++ app/src/main/java/com/example/messenger/ui/login/LoginActivity.kt
@@ -1,15 +1,18 @@
-package com.example.messenger.ui.sample
+/*
+ * LoginActivity.kt 2020. 6. 10
+ *
+ * Copyright 2020 BasicBug. All rights Reserved.
+ *
+ */
+
+package com.example.messenger.ui.login
 
 import android.content.Intent
 import android.os.Bundle
-import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
-import com.example.messenger.NaverLoginManager
 import com.example.messenger.R
 import com.example.messenger.databinding.ActivityLoginBinding
-import com.example.messenger.ui.sample.helper.LoginActionHelper
-import com.nhn.android.naverlogin.ui.view.OAuthLoginButton
 
 /**
  * @author bsgreentea
@@ -17,11 +20,8 @@ import com.nhn.android.naverlogin.ui.view.OAuthLoginButton
 
 class LoginActivity : AppCompatActivity() {
 
-    private lateinit var loginButton: OAuthLoginButton
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_login)
 
         val binding: ActivityLoginBinding =
             DataBindingUtil.setContentView(this, R.layout.activity_login)
@@ -29,20 +29,9 @@ class LoginActivity : AppCompatActivity() {
         binding.loginActionHelper = LoginActionHelper()
         binding.context = this
 
-        initView()
     }
-
-    private fun initView() {
-        loginButton = findViewById(R.id.button_login)
-        loginButton.setOAuthLoginHandler(NaverLoginManager)
-    }
-
     // for the test
     override fun startActivityForResult(intent: Intent?, requestCode: Int) {
         super.startActivityForResult(intent, requestCode)
     }
-
-//    fun login() {
-//        loginInstance.startOauthLoginActivity(this@LoginActivity, naverLoginHandler)
-//    }
 }
