+++ app/src/main/res/layout/activity_start.xml
@@ -28,7 +28,7 @@
             android:id="@+id/bottom_navigation"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            app:setOnListener="@{helper}"
+            app:setOnListenerFromHelper="@{helper}"
             android:background="@color/bottomBackground"
             app:itemIconTint="@color/bottomIconText"
             app:itemTextColor="@color/bottomIconText"
